Use "dmypy daemon" and "dmypy check"
Instead of "restart" and "run" with start the daemon with workspace
configuration and emit "dmypy check" commands upon "lint" request.

The issue with previous approach is that the daemon might be busy (e.g.
if it's processing some other files in the workspace) while the status
is checked; in which case (status returns non-zero) we'd restart it,
thus spawning a new daemon while we might rather wait.

The daemon is stopped through an atexit hook.