Avoid restarting the daemon if is it busy
This can happen in case the file being saved is already being processed
by the daemon or when many files need to be checked within a close
time frame. Instead of restarting the daemon, which would make the
situation worse, do not run mypy but return a meaningful message to
client, with Warning severity.