Avoid restarting the daemon
`dmypy run` is said to handle this when needed and restarting the daemon
ourselves may lead to redundant processes being spawn. Instead, if the
"status" is unexpected, we either check for the presence of the status
file (and ignore the error if missing) or we raise a warning, assuming
next `dmypy run` call will fix the issue.